requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# 各自环境所需的不同依赖, 通过 sites/${env}/deploy/requirements_env.txt得到
-r requirements_env.txt

bk_resource==0.4.10
blueapps>=4.10.0,<5

# web server
gunicorn==21.2.0

# django
django>=3,<4
django-filter==2.4.0
django-cors-headers==3.7.0
django-cors-middleware==1.3.1
django-jsonfield-backport==1.0.3
djangorestframework==3.12.4
drf-yasg==1.20.0
drf_spectacular==0.12.0
django-environ==0.10.0
mysqlclient==2.1.1

# staticfiles
whitenoise==5.2.0

# redis
redis==3.2.0
redis-py-cluster==2.1.0
django-redis==5.2.0

# celery
celery==4.4.7
django-celery-beat==2.2.0
django-celery-results==2.0.1

# iam
bk-iam==1.1.19

# ipython
ipython==7.16.3

# pyinstrument
pyinstrument==3.4.2

# json log
JSON-log-formatter==0.4.0

tencentcloud-sdk-python
